Issue #2226 "Support for SocketOptions.SO_REUSEPORT" (https://github.com/eclipse-ee4j/glassfish-grizzly/issues/2226)
Supports SO_REUSEPORT.
SO_REUSEPORT(reusePort) is supported with interfaces very similar to the set of SO_REUSEADDR(reuseAddress).
The default value of reusePort is false.

Since SO_REUSEPORT is not supported on all platforms, StandardSocketOptions.SO_REUSEPORT is set only in environments where it is possible through a logic check for SO_REUSEPORT support that is executed only once.