Issue #2228 "Grizzly NPE in DefaultFilterChain, this.serverRequest is null" (https://github.com/eclipse-ee4j/glassfish-grizzly/issues/2228) Issue #2214 "NPE when closing OutputBuffer" (https://github.com/eclipse-ee4j/glassfish-grizzly/issues/2214) + Added FINE or FINEST log messages for HttpServerFilter's debugging. + If an exception occurs during the Suspended Request processing, it is handled gracefully on its own without being thrown.
+ Improvements to the use of recycled InputBuffer and OutputBuffer and concurrent access to them.
+ Handle exceptions more elegantly when NIOInputStream, NIOOuputStream, NIOReader, and NIOWriter are not initialized (or recycled).
When a request and response are completed and recycled, the associated ServerInputBuffer and ServerOutputBuffer are also recycled.
However, if another thread references the InputBuffer and OutputBuffer, an NPE may occur when using these recycled instances.
We consider the possibility of unintended usage after recycling, and the possibility of concurrent calls to recycle and use.
I did not modify InputBuffer and OutputBuffer to be completely thread-safe; Improvements have been made only to the areas mentioned in the issues above.

+ Added test cases that can cause NPEs that elegantly handle exceptions.